+++ Makefile
@@ -1,5 +1,5 @@
 fmt:
-	clang-format -i -style=Google $(wildcard *.h */*.h *.c */*.c libc/include/*.h libc/src/*.c)
+	clang-format -i -style=Google $(wildcard *.h */*.h *.c */*.c libc/*.h libc/src/*.c)
 	git diff --exit-code
 
 .PHONY: fmt
