+++ libc/entry.h
@@ -1,6 +1,11 @@
 #ifndef CKB_C_STDLIB_ENTRY_H_
 #define CKB_C_STDLIB_ENTRY_H_
 
+#ifndef CKB_DECLARATION_ONLY
+#ifndef __IMPL_INCLUDED__
+#define __IMPL_INCLUDED__
+#include "src/impl.c"
+
 #ifndef __SHARED_LIBRARY__
 __attribute__((visibility("default"))) __attribute__((naked)) void _start() {
   asm volatile(
@@ -21,5 +26,7 @@ __attribute__((visibility("default"))) __attribute__((naked)) void _start() {
       "ecall");
 }
 #endif /* __SHARED_LIBRARY__ */
+#endif /*__IMPL_INCLUDED__ */
+#endif /* CKB_DECLARATION_ONLY*/
 
 #endif /* CKB_C_STDLIB_ENTRY_H_ */
