--- libc/include/internal/types.h
@@ -1,22 +1,0 @@
-#ifndef HEADER_CKB_C_STDLIB_INTERNAL_TYPES_H_
-#define HEADER_CKB_C_STDLIB_INTERNAL_TYPES_H_
-
-typedef unsigned long size_t;
-typedef signed long ssize_t;
-
-typedef unsigned long uintptr_t;
-typedef signed long intptr_t;
-
-typedef unsigned char uint8_t;
-typedef signed char int8_t;
-
-typedef unsigned short uint16_t;
-typedef signed short int16_t;
-
-typedef unsigned int uint32_t;
-typedef signed int int32_t;
-
-typedef unsigned long uint64_t;
-typedef signed long int64_t;
-
-#endif
