--- libc/include/memory.h
@@ -1,2 +1,0 @@
-
-#include <string.h>
