--- libc/include/stdbool.h
@@ -1,8 +1,0 @@
-#ifndef HEADER_CKB_C_STDLIB_STDBOOL_H_
-#define HEADER_CKB_C_STDLIB_STDBOOL_H_
-
-#define true 1
-#define false 0
-#define bool _Bool
-
-#endif
