--- libc/include/stddef.h
@@ -1,7 +1,0 @@
-#ifndef HEADER_CKB_C_STDLIB_STDDEF_H_
-#define HEADER_CKB_C_STDLIB_STDDEF_H_
-#include "internal/types.h"
-
-#define NULL ((void*)0)
-
-#endif
