--- libc/include/stdio.h
@@ -1,8 +1,0 @@
-
-#ifndef HEADER_CKB_C_STDLIB_STDIO_H_
-#define HEADER_CKB_C_STDLIB_STDIO_H_
-
-int printf(const char* format, ...);
-int ckb_debug(const char* s);
-
-#endif
