--- libc/include/string.h
@@ -1,16 +1,0 @@
-#ifndef HEADER_CKB_C_STDLIB_STRING_H_
-#define HEADER_CKB_C_STDLIB_STRING_H_
-
-#include "internal/types.h"
-
-void *memset(void *dest, int c, size_t n);
-
-void *memcpy(void *restrict dest, const void *restrict src, size_t n);
-void *memmove(void *dest, const void *src, size_t n);
-int memcmp(const void *vl, const void *vr, size_t n);
-char *strcpy(char *restrict d, const char *restrict s);
-size_t strlen(const char *s);
-int strcmp(const char *l, const char *r);
-char *strstr(const char *, const char *);
-
-#endif
