+++ libc/memory.h
@@ -1,2 +1,3 @@
+
 #include <entry.h>
 #include <string.h>
