--- libc/src/memory.c
@@ -1,2 +1,0 @@
-
-#include "../memory.h"
