--- libc/src/stdio.c
@@ -1,2 +1,0 @@
-
-#include "../stdio.h"
