--- libc/src/stdlib.c
@@ -1,2 +1,0 @@
-
-#include "../stdlib.h"
