--- libc/src/string.c
@@ -1,2 +1,0 @@
-
-#include "../string.h"
