+++ libc/stdbool.h
@@ -1,6 +1,5 @@
 #ifndef CKB_C_STDLIB_STDBOOL_H_
 #define CKB_C_STDLIB_STDBOOL_H_
-
 #include <entry.h>
 
 #define true 1
