+++ libc/stddef.h
@@ -4,7 +4,6 @@
 #include <entry.h>
 
 #include "internal/types.h"
-
 #define NULL ((void*)0)
 
 #endif /* CKB_C_STDLIB_STDDEF_H_ */
